feat: Changed Deck and DeckRow components as asked on code review
